user can get a trip by its id
