+++ api/controllers/trips.controllers.js
@@ -1,4 +1,8 @@
-const { selectTrips, postTrip } = require("../models/trips.models");
+const {
+  selectTrips,
+  postTrip,
+  selectSingleTrip,
+} = require("../models/trips.models");
 
 exports.getTrips = (req, res, next) => {
   const { username } = req.query;
@@ -15,6 +19,19 @@ exports.getTrips = (req, res, next) => {
 exports.addNewTrip = (req, res, next) => {
   const newTrip = req.body;
   postTrip(newTrip)
+    .then((trip) => {
+      res.status(200).send({ trip });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
+
+exports.getSingleTrip = (req, res, next) => {
+  const { trip_id } = req.params;
+  const { username } = req.query;
+
+  selectSingleTrip(trip_id, username)
     .then((trip) => {
       res.status(200).send({ trip });
     })
